4.15-4.22

+++ part4/BlogList/controllers/blog.js
@@ -1,9 +1,12 @@
 const express = require('express')
 const blogRouter = express.Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+require('dotenv').config()
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -12,24 +15,38 @@ blogRouter.post('/', async (request, response) => {
     return response.status(400).json({ error: 'title or url missing' })
   }
 
+  const user = request.user
+
   const blog = new Blog({
     title: request.body.title,
     author: request.body.author,
     url: request.body.url,
     likes: request.body.likes || 0,
+    user: user.id,
   })
 
   const savedBlog = await blog.save()
-  response.status(200).json(savedBlog)
+  user.blogs = user.blogs.concat(savedBlog.id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndDelete(request.params.id)
+  const blog = await Blog.findById(request.params.id)
   if (!blog) {
-    response.status(400).json({ error: 'blog not found' })
-  } else {
-    response.status(204).json(blog)
+    return response.status(404).json({ error: 'blog not found' })
   }
+
+  const user = request.user
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized' })
+  }
+
+  await Blog.findByIdAndDelete(request.params.id)
+
+  response.status(204).json({ message: 'blog deleted' })
 })
 
 blogRouter.put('/:id', async (request, response) => {
